pyproject.toml
[project]
name = "doctors-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "openai>=2.15.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "streamlit>=1.53.0",
    "uvicorn>=0.40.0",
]
